dds-agent-framework/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dds-tools"
version = "0.1.0"
description = "CLI tools for AI-assisted DDS development using RTI DDS Spy"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "DDS Agent Framework Team" }
]
keywords = ["dds", "rti", "connext", "middleware", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-timeout>=2.0",
]

[project.scripts]
dds-spy-wrapper = "dds_tools.cli.spy_wrapper:main"
dds-process-monitor = "dds_tools.cli.process_monitor:main"
dds-sample-compare = "dds_tools.cli.sample_compare:main"
dds-test-harness = "dds_tools.cli.test_harness:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dds_tools"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 10  # Default 10 second timeout per test

